car增删改查完成

+++ src/main/java/com/liao/insurance/entity/Car.java
@@ -41,7 +41,7 @@ public class Car implements Serializable {
     /**
      * 估价
      */
-    private BigDecimal evaluatePrice;
+    private Double evaluatePrice;
 
 
 }
